pyproject.toml
[project]
name = "patchfinder"
version = "0.1.0"
description = "Plug-and-play document readability and extraction toolkit for VLMs."
authors = [{ name = "teilomillet", email = "teilomillet@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
    "numpy>=2.2.2",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "pytest>=7.0.0",
    "scipy>=1.11.0",
    "torchvision>=0.15.0",
    "fire>=0.5.0"
]

[project.scripts]
patchfinder = "patchfinder.cli:main"

[build-system]
requires = ["uv>=0.1.0", "setuptools>=65.0.0"]
build-backend = "setuptools.build_meta"
